requirement.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
cryptography==42.0.5
Django==5.0.2
django-anymail==10.2
django-crispy-forms==2.1
django-filter==23.5
django-localflavor==4.0
idna==3.6
pillow==10.2.0
pycparser==2.21
python-stdnum==1.19
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
